# docs/api_surface/OpenAPI.yaml
openapi: 3.0.3
info:
  title: Campus Resource Hub API
  version: 1.0.0
  description: |
    RESTful API for the Campus Resource Hub - A university resource booking and management system.
    
    **Implementation Status:**
    - ✅ existing: Backend implemented and functional
    - ⚠️ partial: Partially implemented
    - ❌ missing: Not yet implemented
    
    **Authentication:** Session-based cookies (Flask-Login)
    
    **Base URL:** `/api`
  contact:
    name: Campus Resource Hub Team
    email: support@campusresourcehub.edu
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://production-domain.com/api
    description: Production server (TBD)
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask-Login session cookie for authentication
    
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF protection token required for all state-changing operations (POST, PUT, PATCH, DELETE).
        Obtain token from GET /api/auth/csrf-token endpoint.
        Token lifetime: 1 hour
        Alternative header name: X-CSRFToken
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [student, staff, admin]
        department:
          type: string
        status:
          type: string
          enum: [active, pending, suspended]
        created_at:
          type: string
          format: date-time
      required: [id, name, email, role]
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
  responses:
    UnauthorizedError:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      x-status: existing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                role:
                  type: string
                  enum: [student, staff, admin]
                  default: student
                department:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error or email exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      x-status: existing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                remember_me:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      x-status: existing
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Logout successful
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      x-status: existing
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    patch:
      tags: [Authentication]
      summary: Update current user profile
      x-status: existing
      security:
        - SessionCookie: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                department:
                  type: string
                profile_image:
                  type: string
      responses:
        '200':
          description: Profile updated
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change password
      x-status: existing
      security:
        - SessionCookie: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Password changed successfully
...
  /auth/check-email:
    post:
      tags: [Authentication]
      summary: Check email availability
      x-status: existing
      description: Check if an email address is already registered
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Email availability status
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
...